agent/requirements.txt
livekit-agents[google,images]>=1.2.0
livekit-plugins-noise-cancellation~=0.2
python-dotenv
asyncio
# google-genai  # Commented out - web search disabled (also reduces cold start time)
